+++ api/views/note/view.py
@@ -1,41 +1,29 @@
-import api.views.note.models as note_models
-import api.views.notes.models as notes_models
-import api.views.dataObjects as dataObjects
-import api.errors as api_errors
-import json
 import ujson
-from api.validator import GetNoteSchema, NewNoteSchema
+import api.views.note.models as note_models
+import api.views.dataObjects as dataObjects
+from api.helpers.check_json_data import check_json_data
+
+from api.validator import GetNoteSchema, GetIpfsNoteSchema, NewNoteSchema
 from aiohttp import web
 from api.route_table import routes
 
 
-def check_json_data(fn):
-    async def wrapper(self):
-        request = self.request
-        request_ = request
-        if isinstance(request, web.View):
-            request_ = request.request
-        assert isinstance(request_, web.Request)
-        try:
-            data = await request.json()
-        except json.decoder.JSONDecodeError:
-            return web.json_response(api_errors.JSON_PARSE_ERROR, status=400)
-        return await fn(self=self, data=data)
-
-    return wrapper
-
-
 @routes.view('/note')
-class Book(web.View):
-
-    @check_json_data
-    async def get(self, data):
+class Note(web.View):
+    @routes.get('/note/{book}/{hash}')
+    async def get(self):
         import api.views.book.models as book_models
+        book = self.match_info['book']
+        hash = self.match_info['hash']
+        data = {
+            'book': book,
+            'hash': hash
+        }
         schema = GetNoteSchema()
         result = schema.load(data)
         if len(result.errors) > 0:
             return web.json_response({'errors': result.errors}, status=400)
-        book_name = data.get('book_name')
+        book_name = data.get('book')
         blockchain_hash = data.get('hash')
         book = book_models.Book(name=book_name)
         note = note_models.Note(book=book, blockchain_hash=blockchain_hash)
@@ -44,6 +32,20 @@ class Book(web.View):
             return web.json_response(text=ujson.dumps(note), status=200)
         else:
             return web.json_response(status=404)
+
+    @routes.get('/note/{ipfs_hash}')
+    async def get_ipfs(self):
+        data = {
+            'ipfs_hash': self.match_info['ipfs_hash']
+        }
+        schema = GetIpfsNoteSchema()
+        result = schema.load(data)
+        if len(result.errors) > 0:
+            return web.json_response({'errors': result.errors}, status=400)
+        ipfs_hash = data.get('ipfs_hash')
+        note = note_models.Note(ipfs_hash=ipfs_hash)
+        ipfs_note = await note.get_ipfs()
+        return web.json_response(text=ujson.dumps(ujson.loads(ipfs_note)))
 
     @check_json_data
     async def post(self, data):
@@ -56,9 +58,10 @@ class Book(web.View):
         book_password = data.get('book_password')
         header = data.get('header')
         content = data.get('content')
-        note = note_models.Note(book=book_models.Book(name=book_name, password=book_password), blockchain_hash="")
-        _note = dataObjects.Note(header=header, content=content, date=1)
-        if await note.add(note=_note):
+        book = book_models.Book(name=book_name, password=book_password)
+        note = note_models.Note(book=book)
+        note_body = dataObjects.Note(header=header, content=content)
+        if await note.add(body=note_body):
             return web.json_response(status=200)
         else:
             return web.json_response(status=400)