+++ api/views/notes/view.py
@@ -1,36 +1,20 @@
-import api.views.notes.models as notes_models
-import api.views.dataObjects as dataObjects
-import api.errors as api_errors
-import json
 import ujson
-from api.validator import GetNotesSchema, NewNoteSchema
+import api.views.notes.models as notes_models
+from api.validator import GetNotesSchema
 from aiohttp import web
 from api.route_table import routes
-
-
-def check_json_data(fn):
-    async def wrapper(self):
-        request = self.request
-        request_ = request
-        if isinstance(request, web.View):
-            request_ = request.request
-        assert isinstance(request_, web.Request)
-        try:
-            data = await request.json()
-        except json.decoder.JSONDecodeError:
-            return web.json_response(api_errors.JSON_PARSE_ERROR, status=400)
-        return await fn(self=self, data=data)
-
-    return wrapper
 
 
 @routes.view('/notes')
 class Book(web.View):
 
-    @check_json_data
-    async def get(self, data):
+    @routes.get('/notes/{book_name}')
+    async def get(self):
         import api.views.book.models as book_models
         schema = GetNotesSchema()
+        data = {
+            'book_name': self.match_info['book_name']
+        }
         result = schema.load(data)
         if len(result.errors) > 0:
             return web.json_response({'errors': result.errors}, status=400)